Fix for recent change in data import; Try to exclude tmpnam

+++ src/complex_ex.cpp
@@ -711,7 +711,7 @@ HADT MGL_EXPORT mgl_find_roots_txt_c(const char *func, const char *vars, HCDT in
 	mglEqTxT par;
 	par.var=vars;	par.FillCmplx(func);
 	size_t n = par.str.size();
-	if(ini->GetNx()!=n)	return 0;
+	if(ini->GetNx()!=long(n))	return 0;
 	mreal *xx = new mreal[2*n];
 	mglDataC *res = new mglDataC(ini);
 	for(long j=0;j<ini->GetNy()*ini->GetNz();j++)
